* Note that unlink sets $! on failure.
The docs to unlink didn't explicitly note that it set $! on
failure, unlike the docs on some other system calls do.

While I was in there, I cleansed the entry a little and
added an example of unlinking files one-by-one to find
the ones that fail.

Modern Perl fix: let's call a glob a glob() and not a <*>.

This problem was noted on the Perl Beginner's list:

http://www.nntp.perl.org/group/perl.beginners/2009/11/msg110062.html