fix logic for a workaround in POSIX.xs
There is the assignment "RETVAL = 0" to silence the compiler in case
it incorrectly reports return of an undefined value.
This assignment should be present whenever *any* of the symbols is
not defined.  But the code was written so that the assignment was
compiled only when *none* of the symbols was defined.