[perl #71076] sort with active sub (5.10 regression)
One of the tests in sort.t causes a bus error (or sometimes ‘Undefined
subroutine called’) if run multiple times. This is because sort
decreases the refcount of an active sub used as a comparison routine.

Ironically enough, this test was added by the very change that broke it
(25953/9850bf2).