Export PL_curinterp symbol for MULTIPLICITY without USE_ITHREADS
This is necessary for XS extensions that define PERL_CORE.  In that
situation PERL_GET_CONTEXT will resolve to PL_curinterp, which is
normally not exported (extensions call Perl_Gcurinterp_ptr() to get a
pointer to PL_curinterp instead). With USE_ITHREADS defined
PERL_GET_CONTEXT will expand to Perl_get_context() even inside the
core because the context needs to be fetched from threadlocal storage.