* fix a perlop fix from debian: http://bugs.debian.org/514814
The original said

     C<< <I<filehandle>> >>

The italic is probably meant for "filehandle", which is a stand-in
for a concrete filehandle. The italic should not apply to the
syntax for the line input operator.

The fix turned it inside out and brought back the ugly interior
sequences:

     I< C< E<lt>filehandleE<gt> > >

It's much better without worrying about the italic face.