Correct \p{print} to not match LINE SEPARATOR nor PARAGRAPH SEPARATOR
The Unicode Standard defines (as a recommendation) that Print be based on
graphical characters and blank characters (minus controls).  Perl's has been
based on space rather than blank. The only practical effect this has is that
Perl erroneously matches the LINE SEPARATOR and PARAGRAPH SEPARATOR, which
clearly are not printable characters.

Signed-off-by: Abigail <abigail@abigail.be>