PATCH: minor typo cleanup of pod/ directory
These are all in the pod/ directory, and only the first is a code fix.
There was also a single lingering ISO 8859-1 encoding that missed the
UTF-8 upconvert.  The rest are cleanups for typos, some of which seem
to have been around for a rather long time: spelling errors, incorrect
possessives, and extra, missing, or duplicated words.

If you actually read through, I bet you'll realize what sparked this. :)

--tom

Signed-off-by: Abigail <abigail@abigail.be>