Correct errors in the use of multiple targets, which could break parallel make.
Where a single script (or program) generates multiple files, the Makefile rule
was written with those multiple files as targets on the same line. This would
be correct if that syntax meant "these $n things are built by these commands".

However it doesn't - it means "to build any of these targets, run these
commands", and hence to run them once for each of the targets. This can result
in race conditions with parallel makes, with files being trampled on and
strange failure modes.

Hence the correct style is to give one of the targets on the rule, and rules
that make all the other targets that are generated depend on that target.