Back out the {ENTER,LEAVE}_with_name("sub") part of d343c3ef45381352 for now.
It's conflicting with an established pattern in XS code, working around the
inability of the XS SCOPE: keyword to actually provide anything useful.
The minor amount of extra debugging it gives here is not worth the external
trouble it causes.

Revisit this once we can provide a meaningful option to disable the
ENTER/LEAVE around XSUBs.