Correct VMS-specific handling of $! in Perl_magic_get.
0097b436152452e403cc71b4f1a1cfd30ec0ba1a had introduced a test failure
on VMS because it only set the POK flag in the non-VMS code path,
which was an easy mistake to make because accumulated patches had
over time made the #ifdef jungle way more tangled than it needed to
be.  There is really only one line that needs to be VMS-specific.