Fix for #71506: work around possible gcc bug
When memcpy() is used on a long double pointer with gcc 4.4 in some
cases it seems to treat it as a long double assignment, copying only
the first 12 bytes.  Use unions so the types we're copying into the
pack output or from the unpack input are unsigned char[], to avoid the
apparent bug.