stop op/groups.t skipping on Linux
My linux (with selinux) includes a "context=" field as part of the
output of "id -a". Make the extraction of the "groups=(.+)" entry
of "id -a" more robust in the presence of unpredictable (in name and
ordering) fields, by splitting on fieldname=(.*), then looking for a field
name of 'groups'.