Add some symbols from version::vxs to the default share
It appears that the core version.pm works with Safe, but that the
CPAN-upgraded version does not. This will need more investigation
and test coverage, but for now, share the minimal set of symbols
needed to load version.pm in a Safe compartment. This covers
only the XS version of version.pm, not the pure perl alternative.