PATCH: memory leak introduced in 5.12.0
There is a small possibility of a memory leak in toke.c when there is a
deprecated character in the name in a \N{...} construct, and the Perl is
embedded or something like that so that memory isn't freed up when it
exits.  This patch avoids the creation of a new scalar, and gives a
better error message besides.