deprecation warning on deprecated-in-core .pl libs
This is a mandatory warning, not obeying -X or lexical warning bits.
To make it behave like an internal deprecation warning would be a lot
of hassle to do in pure Perl.

The warning is modelled on that supplied by deprecate.pm for
deprecated-in-core .pm libraries.  It points to the specific CPAN
distribution that contains the .pl libraries.  The CPAN version, of
course, does not generate the warning.