Deal with "\c{", and its kin
make regen is needed

This patch forbids non-ascii following the "\c".  It also terminates for
"\c{" with a message to contact p5p if there is need for continuing its
current definition.  And if the character following the "\c" causes the
result to not be a control character, a warning is issued.  This is
currently 'deprecated', which by default is turned on.  This can easily
be changed later.

This patch is the initial patch.  It does not do any fancy showing the
context where the problematic construct occurs.  This can be added
later.

It gathers the 3 occurrences of evaluating \c and puts them in one
common routine.