follow up fix for fd69380d5d5b95ef16e2521cf4251b34ee0ce151
The original fix for tied elements losing magic had a bug that was masked
by a bool casting issue. Once the casting was fixed, the bug surfaced:
elements of @+ lost their values when returned from a sub. By removing the
TEMP flag from the regdatum PVLV, we force it to be copied when returned.