Fix Module::Build::Platform::cygwin runtime error
This addresses runtime errors in Module::Build::Platform::cygwin
as reported by Jesse Vincent.  The changes are backported from
the Module::Build repository and the version number has been
slightly incremented to indicate a change from the version on CPAN.

This hopefully addresses test errors in t/actions/installdeps.t, but
the error seems specific to building perl on cygwin and has not been
reported when Module::Build was tested on cygwin by CPAN Testers.