Change perlgpl.pod to GPL 1 to match README
The README and Copying files are specific that Perl is licensed
under Artistic or GPL 1 (or later).  For whatever reason, perlgpl.pod
was created with GPL 2.  This commit changes it back to GPL 1 for
consistency with the actual license.

It also standardizes the introduction to both perlgpl and perlartistic
for consistent language explaining the dual license.